basic refactoring and log message truncate functionality
